Added decision node O20.
I factored the code that adds the body to the response into it's own
function, we're now using that function during decision node O18 to
add the body to the response at that point. We do this so that we can
check to see if the body is present at node O20.

If we bail out early by returning a response code 200, this same
method is used at that point to add the body to the response.

We do not calculate the size of the response or remove it if we're
handling a HEAD request; these operations are delegated to Ring and
Jetty.