Coded up decision node P11.
Altered the handler functions to return a map instead of only the body
content. In this way headers may be manually attached to the response
map, the map returned by the handler is merged with the response
generated by Bishop: all values are merged into one larger map except
the body which is concatenated.

Altered decision node P11 to generate the response body, this prevents
all PUT requests from missing content bodies.