Altered the way node N11 works to better handle create failure.
Node N11 assumed that the creation of a new object was always
successful, this is a mistaken assumption. The logic in this node now
checks to see if the handler function returned it's own status code,
if that is the case then that code is used instead of a 303; if we're
not returning a 303 then the "Location" header is also removed.

This will let you catch exceptions in your handler methods and return
your own status and body content, for instance something like...

  {:status 409 :body "Change would introduce conflict"}