Fixed problem with the root route "/" not matching wildcard.
The root route tokenizes to an empty vector ("[]") and the route logic
wouldn't match that to the wildcard route ("[*]"), this would cause a
status 404 to be returned to the client. I added additional logic to
ensure that an empty URI would be mapped to the wildcard route.

After revisiting the [Webmaching dispatch
documentation](http://wiki.basho.com/Webmachine-Dispatching.html), it
looks like this is the only way to match the root of the URI
namespace; that is this is the expected behavior.

I also added some tests to catch any future bugs in the routing code.

This was reported on [issue
#5](https://github.com/tnr-global/bishop/issues/5).