Fixed problem with the character set not being set.
This was reported in issue #9. [0]

The character set on the response content-type was only being set if
the client provided an "Acceptable-Charset" header. The thinking was
that if no character set was provided, none would be set on the
response. This is incorrect, the default character set is ISO-8859-1;
if no character is provided then that is the one used. This is from
the HTTP 1.1 spec[1] and is enforced by Jetty[2].

While ISO-8859-1 is the "official" default character encoding, it is
not the encoding recommended by the W3C. Unsurprisingly, the
recommended encoding is UTF-8[3] and that's the one we should use by
default. The default list of character sets provided by all Bishop
resources has only the one entry, "utf-8" (it was mis-spelled and
that's also been fixed). Since Jetty is setting the encoding if we
don't set it ourselves, I have altered the logic in node E5 to choose
the first listed encoding for the resource if the client is not
providing an "Accept-Charset" header.

The last issue is how to enforce the selected character encoding. The
JVM sets its character encoding to be the same as the host operating
system[4]. The Jetty method "setCharacterEncoding" will be called by
Ring and set to the encoding provided in the response's
"Content-Type". The documentation for Jetty clearly states that this
will set the encoding for the data sent to the client[5]. With this
change, the encoding will not only be set in the header but the
correct output will also be sent to the client.

[0]: https://github.com/tnr-global/bishop/issues/9
[1]: http://www.w3.org/International/O-HTTP-charset.en.php
[2]: http://docs.codehaus.org/display/JETTY/International+Characters+and+Character+Encodings
[3]: http://www.w3.org/International/questions/qa-choosing-encodings
[4]: http://docs.oracle.com/javase/6/docs/technotes/guides/jps/spec/docflavor.fm2.html
[5]: http://api.dpml.net/jetty/5.1.6/javax/servlet/ServletResponse.html#setCharacterEncoding(java.lang.String)