Started work on allowing responses from all callback functions.
In the current release a callback function may return true, a map with
the response body or false. I'm working on changing this so that these
functions may return the following:

*  True
*  False
*  A map with the response map, implies True
*  A sequence, the first item being a True or False and the remainder
   containing the response map

This will allow the return of a response map even when the callback
evaluates to false. In all cases the response map is merged with the
response built up so far; if the callback's response map contains a
Content-Type header this will be preferenced over the value in the
accumulated response.