Fixed issue where 406 response lack a body.
When a 406 response is sent to the client, it should contain a list of
one or more acceptable media types in the body. Bishop will now return
the first most likely acceptable character type based on the following
algorithm.

1.  If the response has been calculated, use that content type
2.  If a content-type is negotiated, use that
3.  Use the first content type provided by the resource

An acceptable character set is chosen in the same manner, by first
inspecting the response, then the negotiated character set and,
lastly, choosing the first character set provided by the resource.

This logic is used in an attempt to provide the most likely
content. That is, if the client has an acceptable content type but not
character set, the suggested media type will be for their chosen
content type with a known good character set.