Fixed a bug in the handling of empty "Accept" headers per issue #16.
I looked through the Webmachine decision tree and an empty "Accept"
header should be treated as if the client finds no content types
acceptable; a 406 "Not Acceptable" should be returned.

The root cause of this bug was in decision node C4, instead of
checking for an empty sequence in the "Accept" header we were checking
for an nil or false. Since the headers are always returned as a
sequence, an empty "Accept" header would be represented by the empty
list. We now check and handle the empty list correctly in this case.