Fixed issue #8 where resources that returned a String would fail.
In the README file, we have an example of a resource that simply
returns a String with the expectation that this will become the body
of the response. Because of the way the resource response is merged
with the response built by Bishop, this type of resource would
fail.

This has now been fixed, if a resource returns a map then that map is
merged with the Bishop response. If the resource returns anything
else, it is placed in the :body key of the response. This will let a
resource function return both a simple String or something that Ring
knows how to handle, like an InputStream or File.

I also incremented the version number for release.